Fix SR bfloat16 NaN issue on CPU

+++ src/sr_module.py
@@ -103,7 +103,8 @@ def load_sr_model(device: str = "cpu") -> "ESPCN | None":
         model = ESPCN(in_channels=3, upscale_factor=2)
         state = torch.load(SR_CKPT, map_location=device)
         model.load_state_dict(state)
-        model.to(device).eval()
+        # Ensure model is float32 to prevent bfloat16 silent NaNs on CPU
+        model.to(device).to(torch.float32).eval()
         _sr_model = model
         return model
     except Exception as e:
